Move assets and `GltfAssetLabel` to new modules on `bevy_gltf`
