+++ crates/bevy_gltf/src/assets/extras.rs
@@ -1,0 +1,49 @@
+use bevy_ecs::{component::Component, reflect::ReflectComponent};
+use bevy_reflect::{prelude::ReflectDefault, Reflect};
+
+/// Additional untyped data that can be present on most glTF types at the primitive level.
+///
+/// See [the relevant glTF specification section](https://registry.khronos.org/glTF/specs/2.0/glTF-2.0.html#reference-extras).
+#[derive(Clone, Debug, Reflect, Default, Component)]
+#[reflect(Component, Default, Debug)]
+pub struct GltfExtras {
+    /// Content of the extra data.
+    pub value: String,
+}
+
+/// Additional untyped data that can be present on most glTF types at the scene level.
+///
+/// See [the relevant glTF specification section](https://registry.khronos.org/glTF/specs/2.0/glTF-2.0.html#reference-extras).
+#[derive(Clone, Debug, Reflect, Default, Component)]
+#[reflect(Component, Default, Debug)]
+pub struct GltfSceneExtras {
+    /// Content of the extra data.
+    pub value: String,
+}
+
+/// Additional untyped data that can be present on most glTF types at the mesh level.
+///
+/// See [the relevant glTF specification section](https://registry.khronos.org/glTF/specs/2.0/glTF-2.0.html#reference-extras).
+#[derive(Clone, Debug, Reflect, Default, Component)]
+#[reflect(Component, Default, Debug)]
+pub struct GltfMeshExtras {
+    /// Content of the extra data.
+    pub value: String,
+}
+
+/// Additional untyped data that can be present on most glTF types at the material level.
+///
+/// See [the relevant glTF specification section](https://registry.khronos.org/glTF/specs/2.0/glTF-2.0.html#reference-extras).
+#[derive(Clone, Debug, Reflect, Default, Component)]
+#[reflect(Component, Default, Debug)]
+pub struct GltfMaterialExtras {
+    /// Content of the extra data.
+    pub value: String,
+}
+
+/// The material name of a glTF primitive.
+///
+/// See [the relevant glTF specification section](https://registry.khronos.org/glTF/specs/2.0/glTF-2.0.html#reference-material).
+#[derive(Clone, Debug, Reflect, Default, Component)]
+#[reflect(Component)]
+pub struct GltfMaterialName(pub String);
