+++ crates/bevy_gltf/src/assets/mesh.rs
@@ -1,0 +1,47 @@
+use bevy_asset::Asset;
+use bevy_reflect::TypePath;
+
+use crate::label::GltfAssetLabel;
+
+use super::{GltfExtras, GltfPrimitive};
+
+/// A glTF mesh, which may consist of multiple [`GltfPrimitives`](GltfPrimitive)
+/// and an optional [`GltfExtras`].
+///
+/// See [the relevant glTF specification section](https://registry.khronos.org/glTF/specs/2.0/glTF-2.0.html#reference-mesh).
+#[derive(Asset, Debug, Clone, TypePath)]
+pub struct GltfMesh {
+    /// Index of the mesh inside the scene
+    pub index: usize,
+    /// Computed name for a mesh - either a user defined mesh name from gLTF or a generated name from index
+    pub name: String,
+    /// Primitives of the glTF mesh.
+    pub primitives: Vec<GltfPrimitive>,
+    /// Additional data.
+    pub extras: Option<GltfExtras>,
+}
+
+impl GltfMesh {
+    /// Create a mesh extracting name and index from glTF def
+    pub fn new(
+        mesh: &gltf::Mesh,
+        primitives: Vec<GltfPrimitive>,
+        extras: Option<GltfExtras>,
+    ) -> Self {
+        Self {
+            index: mesh.index(),
+            name: if let Some(name) = mesh.name() {
+                name.to_string()
+            } else {
+                format!("GltfMesh{}", mesh.index())
+            },
+            primitives,
+            extras,
+        }
+    }
+
+    /// Subasset label for this mesh within the gLTF parent asset.
+    pub fn asset_label(&self) -> GltfAssetLabel {
+        GltfAssetLabel::Mesh(self.index)
+    }
+}
