+++ crates/bevy_gltf/src/assets/mod.rs
@@ -1,0 +1,52 @@
+//! Representation of assets present in a glTF file
+
+pub mod extras;
+pub mod mesh;
+pub mod node;
+pub mod primitive;
+pub mod skin;
+
+#[cfg(feature = "bevy_animation")]
+use bevy_animation::AnimationClip;
+use bevy_asset::{Asset, Handle};
+use bevy_pbr::StandardMaterial;
+use bevy_platform_support::collections::HashMap;
+use bevy_reflect::TypePath;
+use bevy_scene::Scene;
+
+use self::{extras::*, mesh::*, node::*, primitive::*, skin::*};
+
+/// Representation of a loaded glTF file.
+#[derive(Asset, Debug, TypePath)]
+pub struct Gltf {
+    /// All scenes loaded from the glTF file.
+    pub scenes: Vec<Handle<Scene>>,
+    /// Named scenes loaded from the glTF file.
+    pub named_scenes: HashMap<Box<str>, Handle<Scene>>,
+    /// All meshes loaded from the glTF file.
+    pub meshes: Vec<Handle<GltfMesh>>,
+    /// Named meshes loaded from the glTF file.
+    pub named_meshes: HashMap<Box<str>, Handle<GltfMesh>>,
+    /// All materials loaded from the glTF file.
+    pub materials: Vec<Handle<StandardMaterial>>,
+    /// Named materials loaded from the glTF file.
+    pub named_materials: HashMap<Box<str>, Handle<StandardMaterial>>,
+    /// All nodes loaded from the glTF file.
+    pub nodes: Vec<Handle<GltfNode>>,
+    /// Named nodes loaded from the glTF file.
+    pub named_nodes: HashMap<Box<str>, Handle<GltfNode>>,
+    /// All skins loaded from the glTF file.
+    pub skins: Vec<Handle<GltfSkin>>,
+    /// Named skins loaded from the glTF file.
+    pub named_skins: HashMap<Box<str>, Handle<GltfSkin>>,
+    /// Default scene to be displayed.
+    pub default_scene: Option<Handle<Scene>>,
+    /// All animations loaded from the glTF file.
+    #[cfg(feature = "bevy_animation")]
+    pub animations: Vec<Handle<AnimationClip>>,
+    /// Named animations loaded from the glTF file.
+    #[cfg(feature = "bevy_animation")]
+    pub named_animations: HashMap<Box<str>, Handle<AnimationClip>>,
+    /// The gltf root of the gltf asset, see <https://docs.rs/gltf/latest/gltf/struct.Gltf.html>. Only has a value when `GltfLoaderSettings::include_source` is true.
+    pub source: Option<gltf::Gltf>,
+}
