+++ crates/bevy_gltf/src/assets/node.rs
@@ -1,0 +1,74 @@
+use bevy_asset::{Asset, Handle};
+use bevy_reflect::TypePath;
+
+use crate::label::GltfAssetLabel;
+
+use super::{GltfExtras, GltfMesh, GltfSkin};
+
+/// A glTF node with all of its child nodes, its [`GltfMesh`],
+/// [`Transform`](bevy_transform::prelude::Transform), its optional [`GltfSkin`]
+/// and an optional [`GltfExtras`].
+///
+/// See [the relevant glTF specification section](https://registry.khronos.org/glTF/specs/2.0/glTF-2.0.html#reference-node).
+#[derive(Asset, Debug, Clone, TypePath)]
+pub struct GltfNode {
+    /// Index of the node inside the scene
+    pub index: usize,
+    /// Computed name for a node - either a user defined node name from gLTF or a generated name from index
+    pub name: String,
+    /// Direct children of the node.
+    pub children: Vec<Handle<GltfNode>>,
+    /// Mesh of the node.
+    pub mesh: Option<Handle<GltfMesh>>,
+    /// Skin of the node.
+    pub skin: Option<Handle<GltfSkin>>,
+    /// Local transform.
+    pub transform: bevy_transform::prelude::Transform,
+    /// Is this node used as an animation root
+    #[cfg(feature = "bevy_animation")]
+    pub is_animation_root: bool,
+    /// Additional data.
+    pub extras: Option<GltfExtras>,
+}
+
+impl GltfNode {
+    /// Create a node extracting name and index from glTF def
+    pub fn new(
+        node: &gltf::Node,
+        children: Vec<Handle<GltfNode>>,
+        mesh: Option<Handle<GltfMesh>>,
+        transform: bevy_transform::prelude::Transform,
+        skin: Option<Handle<GltfSkin>>,
+        extras: Option<GltfExtras>,
+    ) -> Self {
+        Self {
+            index: node.index(),
+            name: if let Some(name) = node.name() {
+                name.to_string()
+            } else {
+                format!("GltfNode{}", node.index())
+            },
+            children,
+            mesh,
+            transform,
+            skin,
+            #[cfg(feature = "bevy_animation")]
+            is_animation_root: false,
+            extras,
+        }
+    }
+
+    /// Create a node with animation root mark
+    #[cfg(feature = "bevy_animation")]
+    pub fn with_animation_root(self, is_animation_root: bool) -> Self {
+        Self {
+            is_animation_root,
+            ..self
+        }
+    }
+
+    /// Subasset label for this node within the gLTF parent asset.
+    pub fn asset_label(&self) -> GltfAssetLabel {
+        GltfAssetLabel::Node(self.index)
+    }
+}
